requirements.txt
youtube-transcript-api
requests
flask
flask-cors
python-dotenv
